numerous changes;

+++ App.xaml.cs
@@ -29,13 +29,13 @@ namespace ZipImageViewer
         //public const int PreviewCount = 4;
         public static Random Random = new Random();
 
-        internal static ImageSource fa_meh;
-        internal static ImageSource fa_spinner;
-        internal static ImageSource fa_exclamation;
-        internal static ImageSource fa_file;
-        internal static ImageSource fa_folder;
-        internal static ImageSource fa_archive;
-        internal static ImageSource fa_image;
+        public static ImageSource fa_meh { get; private set; }
+        public static ImageSource fa_spinner { get; private set; }
+        public static ImageSource fa_exclamation { get; private set; }
+        public static ImageSource fa_file { get; private set; }
+        public static ImageSource fa_folder { get; private set; }
+        public static ImageSource fa_archive { get; private set; }
+        public static ImageSource fa_image { get; private set; }
 
 
         public static CubicEase CE_EaseIn => (CubicEase)Current.FindResource("CE_EaseIn");
@@ -67,15 +67,17 @@ namespace ZipImageViewer
                 CheckThumbsDB();
 
 #if DEBUG
-                Execute(Table.Thumbs, (table, con) => {
-                    using (var cmd = new SQLiteCommand(con)) {
-                        cmd.CommandText = $@"delete from {table.Name}";
-                        cmd.ExecuteNonQuery();
-                        cmd.CommandText = @"vacuum";
-                        cmd.ExecuteNonQuery();
-                    }
-                    return 0;
-                });
+                if (e.Args?.Length > 0 && e.Args[0] == "-cleandb") {
+                    Execute(Table.Thumbs, (table, con) => {
+                        using (var cmd = new SQLiteCommand(con)) {
+                            cmd.CommandText = $@"delete from {table.Name}";
+                            cmd.ExecuteNonQuery();
+                            cmd.CommandText = @"vacuum";
+                            cmd.ExecuteNonQuery();
+                        }
+                        return 0;
+                    });
+                }
 #endif
 
                 //handle immersion mode change
